adjustments on tailwind config
