+++ package.json
@@ -6,6 +6,7 @@
     "start": "next start"
   },
   "dependencies": {
+    "@headlessui/react": "^1.7.10",
     "framer-motion": "^8.5.2",
     "next": "latest",
     "nookies": "^2.5.2",
@@ -14,8 +15,12 @@
     "react-responsive-carousel": "^3.2.23"
   },
   "devDependencies": {
+    "@types/node": "18.11.18",
+    "@types/react": "18.0.27",
     "autoprefixer": "^10.4.13",
+    "cssnano": "^5.1.14",
     "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.4"
+    "tailwindcss": "^3.2.4",
+    "typescript": "4.9.5"
   }
 }
